+++ src/app/[locale]/(website)/o-nas/hero.tsx
@@ -2,7 +2,7 @@ import { Heading } from "@/src/app/components/heading";
 import { LinkButton } from "@/src/app/components/link-button";
 import { PageWrapper } from "@/src/app/components/page-wrapper";
 import Image from "next/image";
-import team from "../../../../public/assets/team.webp";
+import team from "../../../../../public/assets/team.webp";
 
 export function Hero() {
   return (
