+++ src/app/[locale]/(website)/page.tsx
@@ -3,7 +3,7 @@ import { NextEvent } from "./next-event";
 import { Aktivnosti } from "./aktivnosti";
 import { Reviews } from "./reviews";
 import { TopBlogs } from "./top-blogs";
-import { Footer } from "../components/footer";
+import { Footer } from "@/src/app/components/footer";
 import { Logos } from "./logos";
 import { Discord } from "./discord";
 export default function Page() {
