+++ src/app/[locale]/(website)/partnerstvo/hero.tsx
@@ -3,7 +3,7 @@ import { LinkButton } from "@/src/app/components/link-button";
 import { PageWrapper } from "@/src/app/components/page-wrapper";
 import { Paragraph } from "@/src/app/components/paragraph";
 import Image from "next/image";
-import partners from "../../../../public/assets/partners-image.webp";
+import partners from "../../../../../public/assets/partners-image.webp";
 
 export function Hero() {
   return (
