+++ src/app/[locale]/(website)/partnerstvo/partners-list.tsx
@@ -1,4 +1,3 @@
-import { Heading } from "@/src/app/components/heading";
 import { PageWrapper } from "@/src/app/components/page-wrapper";
 import { Paragraph } from "@/src/app/components/paragraph";
 import Image from "next/image";
