+++ src/app/[locale]/globals.css
